+++ __tests__/lib/services/config-service.test.ts
@@ -1,4 +1,4 @@
-import { ConfigService } from "../../src/lib/server/services/config-service";
+import { ConfigService } from "../../../src/lib/server/services/config-service";
 import { expect, test } from "vitest";
 
 test("ConfigService has getApiUrl method", () => {
